♻️ Extract db instance to break circular dependencies
Sub-modules (connections.ts, users.ts, notifications.ts) were importing
db from index.ts, while index.ts re-exports from those sub-modules.
This created a circular dependency chain.

Solution: Extract the db client instance to a dedicated client.ts file.
Sub-modules now import db from client.ts instead of the barrel file.
External imports via @/lib/db still work via re-export.
