+++ lib/db/connections.ts
@@ -7,7 +7,7 @@
 
 import { eq, desc, and } from "drizzle-orm";
 
-import { db } from "./index";
+import { db } from "./client";
 import {
     connections,
     messages,
